// angular-sample/angular-webapp/client/app/employee/models/vehicle-details.ts
export class VehicleDetails {
    Make: string;
    Model: string;
    Color: string;
    Details: string;
    MOTStartDate: Date;
    MOTEndDate: Date;
    DateIssued: Date;
    ReturnDate: Date;
    FirstRegistered: Date;
    Registration: string;
    EngineCC: string;
    ApprovedCO2: string;
    FuelType: string;
    InsuranceStartDate: Date;
    InsuranceEndDate: Date;
    InsuranceCompany: string;
    PolicyNumber: string;
    InsuranceValue: string;
    Allowance: string;
    EngineCc: string;
    Id: string;
    EmployeeId:string;  
}